MAINT: shifted compute logic to API method
