+++ sucpp/su.cpp
@@ -2,11 +2,8 @@
 #include <fstream>
 #include <string>
 #include <iomanip>
-#include "tree.hpp"
-#include "biom.hpp"
-#include "unifrac.hpp"
+#include "api.hpp"
 #include "cmd.hpp"
-#include <thread>
 
 void usage() {
     std::cout << "usage: ssu -i <biom> -o <out.dm> -m [METHOD] -t <newick> [-n threads] [-a alpha] [--vaw]" << std::endl;
@@ -35,12 +32,6 @@ void usage() {
 void err(std::string msg) {
     std::cerr << "ERROR: " << msg << std::endl << std::endl;
     usage();
-}
-
-// https://stackoverflow.com/a/19841704/19741
-bool is_file_exists(const char *fileName) {
-    std::ifstream infile(fileName);
-        return infile.good();
 }
 
 int main(int argc, char **argv){
@@ -81,15 +72,9 @@ int main(int argc, char **argv){
         err("table filename missing");
         return EXIT_FAILURE;
     }
-    if(!is_file_exists(table_filename.c_str())) {
-        return EXIT_FAILURE;
-    }
 
     if(tree_filename.empty()) {
         err("tree filename missing");
-        return EXIT_FAILURE;
-    }
-    if(!is_file_exists(tree_filename.c_str())) {
         return EXIT_FAILURE;
     }
 
@@ -117,76 +102,16 @@ int main(int argc, char **argv){
         g_unifrac_alpha = atof(gunifrac_arg.c_str());
     }
 
-    if(method_string == "unweighted") 
-        method = su::unweighted;
-    else if(method_string == "weighted_normalized")
-        method = su::weighted_normalized;
-    else if(method_string == "weighted_unnormalized")
-        method = su::weighted_unnormalized;
-    else if(method_string == "generalized")
-        method = su::generalized;
-    else {
-        err("Unknown method");
-        return EXIT_FAILURE;
+    double **dm = NULL;
+    int err;
+    err = su::one_off(table_filename.c_str(), tree_filename.c_str(), method_string.c_str(), 
+                      vaw, g_unifrac_alpha, nthreads, dm);
+    if(err != OK || dm == NULL) {
+        fprintf(stderr, "Compute failed: %d\n", err);
+        exit(EXIT_FAILURE);
     }
 
-    std::ifstream ifs((char*)(tree_filename).c_str());
-    std::string content = std::string(std::istreambuf_iterator<char>(ifs), std::istreambuf_iterator<char>());
     su::biom table = su::biom(table_filename);
-    su::BPTree tree = su::BPTree(content);
-    
-    std::unordered_set<std::string> to_keep(table.obs_ids.begin(), table.obs_ids.end());
-    su::BPTree tree_sheared = tree.shear(to_keep).collapse();
-
-    std::vector<double*> dm_stripes; 
-    std::vector<double*> dm_stripes_total; 
-    dm_stripes.resize((table.n_samples + 1) / 2);
-    dm_stripes_total.resize((table.n_samples + 1) / 2);
-    
-    unsigned int chunksize = dm_stripes.size() / nthreads;
-    unsigned int start = 0;
-    unsigned int end = dm_stripes.size();
-    
-    std::vector<su::task_parameters> tasks;
-    tasks.resize(nthreads);
-
-    std::vector<std::thread> threads(nthreads);
-    for(unsigned int tid = 0; tid < threads.size(); tid++) {
-        tasks[tid].tid = tid;
-        tasks[tid].start = start; // stripe start
-        tasks[tid].stop = start + chunksize;  // stripe end 
-        tasks[tid].n_samples = table.n_samples;
-        tasks[tid].g_unifrac_alpha = g_unifrac_alpha;
-        start = start + chunksize;
-    }
-    // the last thread gets any trailing bits
-    tasks[threads.size() - 1].stop = end;
-    
-    for(unsigned int tid = 0; tid < threads.size(); tid++) {
-        if(vaw)
-            threads[tid] = std::thread(su::unifrac_vaw, 
-                                       std::ref(table),
-                                       std::ref(tree_sheared), 
-                                       method, 
-                                       std::ref(dm_stripes), 
-                                       std::ref(dm_stripes_total), 
-                                       &tasks[tid]);
-        else
-            threads[tid] = std::thread(su::unifrac, 
-                                       std::ref(table),
-                                       std::ref(tree_sheared), 
-                                       method, 
-                                       std::ref(dm_stripes), 
-                                       std::ref(dm_stripes_total), 
-                                       &tasks[tid]);
-    }
-
-    for(unsigned int tid = 0; tid < threads.size(); tid++) {
-        threads[tid].join();
-    }
-    
-    double **dm = su::deconvolute_stripes(dm_stripes, table.n_samples);
-
     std::ofstream output;
     output.open(output_filename);
     
@@ -198,14 +123,6 @@ int main(int argc, char **argv){
         for(unsigned int j = 0; j < table.n_samples; j++)
             output << std::setprecision(16) << "\t" << dm[i][j];
         output << std::endl;
-    }
-    unsigned int n_rotations = (table.n_samples + 1) / 2;
-    for(unsigned int i = 0; i < n_rotations; i++)
-        free(dm_stripes[i]);
-    
-    if(method == su::weighted_normalized || method == su::unweighted) {
-        for(unsigned int i = 0; i < n_rotations; i++)
-            free(dm_stripes_total[i]);
     }
     return EXIT_SUCCESS;
 }
